requirements.txt
pytorch==0.4.0
numpy==1.13.1